fix argparse type of gamma

+++ reinforcement_learning/actor_critic.py
@@ -14,7 +14,7 @@ import torchvision.transforms as T
 
 
 parser = argparse.ArgumentParser(description='PyTorch actor-critic example')
-parser.add_argument('--gamma', type=int, default=0.99, metavar='G',
+parser.add_argument('--gamma', type=float, default=0.99, metavar='G',
                     help='discount factor (default: 0.99)')
 parser.add_argument('--seed', type=int, default=543, metavar='N',
                     help='random seed (default: 1)')
